haskellPackages.aeson_1_5_6_0: Fix build failures due to removed package
GHC release 9.2.1 updated the GHC.Base package to 4.16.0.0. This
version of GHC.Base removed the Data.Semigroup.Option typeclass. This
breaks aeson version 1.5.6.0, which breaks several packages which
depend on it.

This commit patches this version of aeson to remove all references to
Data.Semigroup.Option, as Maybe can (and should) be used instead.

See GHC.Base release notes for details:
https://github.com/ghc/ghc/blob/ghc-9.2.1-release/libraries/base/changelog.md